## [2.1.4-SNAPSHOT] ### Fixed: - Prevent java.lang.NumberFormatException while comparing text with alpha-numeric comparator. As alpha-numeric comparator divides text on numbers and letters - there might be a case that number is too small for Integer. Maximum number value for Integer is 2,147,483,647. For example: in text "ABC80504572795201" number 80504572795201 does not match Integer type and exception is thrown. As a resolution, alpha-numeric comparator now treats numbers as Long. Maximum value for Long is 9,223,372,036,854,775,807 which is a much bigger range.

+++ src/main/java/com/paxar/qps/common/utils/AlphanumericComparator.java
@@ -32,7 +32,7 @@ public class AlphanumericComparator implements Comparator<String> {
 
     private int compareChunks(String chunk1, String chunk2) {
         if (StringUtils.isNumeric(chunk1) && StringUtils.isNumeric(chunk2)) {
-            return Integer.valueOf(chunk1).compareTo(Integer.valueOf(chunk2));
+            return Long.valueOf(chunk1).compareTo(Long.valueOf(chunk2));
         } else if (StringUtils.isBlank(chunk1) && StringUtils.isBlank(chunk2)) {
             return 0;
         } else {
